Promoting the game to the site root and fixing issues with link paths to fix broken build in production
