+++ src/routes/how-to-play/+page.svelte
