+++ src/routes/how-to-play/+page.ts
