+++ src/routes/reduced-motion.ts
